+++ pos/src/main/java/se/kth/iv1350/peripherals/Register.java
@@ -1,12 +1,15 @@
 package se.kth.iv1350.peripherals;
 
+/**
+ * Substitute for external register.
+ */
 public class Register {
     
     /**
      * Changes the amount available in the register by the given amount
      * @param amount The amount to change.
      */
-    public void Update(int amount)
+    public void Update(double amount)
     {
         return;
     }
